--- hopsworks-admin/src/main/webapp/security/protected/admin/jobManagement.xhtml
@@ -1,130 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Changes to this file committed after and not including commit-id: ccc0d2c5f9a5ac661e60e6eaf138de7889928b8b
-  ~ are released under the following license:
-  ~
-  ~ This file is part of Hopsworks
-  ~ Copyright (C) 2018, Logical Clocks AB. All rights reserved
-  ~
-  ~ Hopsworks is free software: you can redistribute it and/or modify it under the terms of
-  ~ the GNU Affero General Public License as published by the Free Software Foundation,
-  ~ either version 3 of the License, or (at your option) any later version.
-  ~
-  ~ Hopsworks is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
-  ~ without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
-  ~ PURPOSE.  See the GNU Affero General Public License for more details.
-  ~
-  ~ You should have received a copy of the GNU Affero General Public License along with this program.
-  ~ If not, see <https://www.gnu.org/licenses/>.
-  ~
-  ~ Changes to this file committed before and including commit-id: ccc0d2c5f9a5ac661e60e6eaf138de7889928b8b
-  ~ are released under the following license:
-  ~
-  ~ Copyright (C) 2013 - 2018, Logical Clocks AB and RISE SICS AB. All rights reserved
-  ~
-  ~ Permission is hereby granted, free of charge, to any person obtaining a copy of this
-  ~ software and associated documentation files (the "Software"), to deal in the Software
-  ~ without restriction, including without limitation the rights to use, copy, modify, merge,
-  ~ publish, distribute, sublicense, and/or sell copies of the Software, and to permit
-  ~ persons to whom the Software is furnished to do so, subject to the following conditions:
-  ~
-  ~ The above copyright notice and this permission notice shall be included in all copies or
-  ~ substantial portions of the Software.
-  ~
-  ~ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS  OR IMPLIED, INCLUDING
-  ~ BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-  ~ NONINFRINGEMENT. IN NO EVENT SHALL  THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
-  ~ DAMAGES OR  OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-  ~ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-  -->
-
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
-  "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml"   
-      xmlns:f="http://java.sun.com/jsf/core"
-      xmlns:h="http://java.sun.com/jsf/html"
-      xmlns:ui="http://java.sun.com/jsf/facelets"
-      xmlns:p="http://primefaces.org/ui"
-      >
-  <f:view contentType="text/html">
-
-    <h:head>
-      <title><ui:insert name="title">Hopsworks</ui:insert></title>
-      <link rel="icon" type="image/png" href="#{resource['images/hops-black.png']}" />
-      <link rel="stylesheet" type="text/css" href="#{resource['/css/main.css']}" />
-      <link rel="stylesheet" type="text/css" href="#{resource['/css/theme.css']}" />
-      <meta http-equiv="refresh" content="#{facesContext.externalContext.sessionMaxInactiveInterval};url=#{request.contextPath}/index.xhtml"/>
-      <link href="//maxcdn.bootstrapcdn.com/font-awesome/4.2.0/css/font-awesome.min.css" rel="stylesheet"/>
-    </h:head>
-
-    <h:body>
-      <ui:insert name="titleBar" >
-        <ui:include src="/security/protected/admin/admintitleBar.xhtml" />
-      </ui:insert>
-
-      <p:layout id="UserMgmLayout" styleClass="admin-main-layout" rendered="#{roleEnforcementPoint.admin}">  
-
-        <p:growl autoUpdate="true" showDetail="true"/>
-        <p:layoutUnit id="centerLayout" position="center">
-
-          <div class="container">
-            <div align="left">
-              <h2>Job Administration</h2>
-            </div>
-            <div align="left">
-              <h:outputText value="Number of Jobs:#{jobAdministration.numberOfJobs}" style="font-weight: bold" />
-            </div>
-            <div class="content">
-
-              <h:form id="jobmodifyForm" >
-                <p:growl id="growl" life="2000"/>
-
-                <p:dataTable id="jobdataTable" var="job" 
-                             value="#{jobAdministration.allJobs}" 
-                             filteredValue="#{jobAdministration.filteredJobs}"
-                             widgetVar="jobsTable"
-                             emptyMessage="No jobs found with given criteria"
-                             rowKey="#{job.appId}"
-                             style="table-layout: auto" 
-                             scrollable="false"
-                             rows="25"
-                             paginator="true"
-                             paginatorTemplate="{CurrentPageReport}  {FirstPageLink} {PreviousPageLink} {PageLinks} {NextPageLink} {LastPageLink} {RowsPerPageDropdown}"
-                             rowsPerPageTemplate="10,25,50,100,250,500">
-                  
-                  <p:column headerText="Id" filterBy="#{job.appId}" sortBy="#{job.appId}" style="width: 17%; word-wrap: break-word;">
-                    <h:outputText value="#{job.appId}" />
-                  </p:column>
-
-                  <p:column id="nameColumn" filterBy="#{job.name}" sortBy="#{job.name}" headerText="Name" style="width: 15%; word-wrap: break-word;" >
-                    <h:outputText value="#{job.name}" />
-                  </p:column>
-                  <p:column id="userColumn" filterBy="#{job.user}" sortBy="#{job.user}" headerText="User" style="width: 15%; word-wrap: break-word;">
-                    <h:outputText value="#{job.user}" />
-                  </p:column>
-                  <p:column id="starttimeColumn" filterBy="#{job.startTime}" sortBy="#{job.startTime}" style="width: 10%;" headerText="Start Time" >
-                    <h:outputText value="#{job.startTime}"/>
-                  </p:column>
-                  <p:column id="finishtimeColumn" filterBy="#{job.finishTime}" sortBy="#{job.finishTime}" style="width: 10%;" headerText="Finish Time" >
-                    <h:outputText rendered="#{job.finishTime != null}" value="#{job.finishTime}"/>  
-                    <h:outputText rendered="#{job.finishTime == null}" value="-"/>     
-                  </p:column>
-                  <p:column id="mstateColumn" filterBy="#{job.state}" sortBy="#{job.state}" style="width: 10%;"  headerText="State"  >
-                    <h:outputText value="#{job.state}"/>
-                  </p:column>
-                  <p:column id="actionJobColumn" headerText="Action" >                                                 
-                    <p:commandButton  action="#{jobAdministration.killJob(job.appId)}" update="@form"  value="Kill"/>
-                  </p:column>                  
-                  <p:column id="statusColumn" headerText="Status" >
-                    <h:outputText value="#{jobAdministration.getError(job.appId)}" />
-                  </p:column>
-                </p:dataTable>                                            
-              </h:form>   
-
-            </div>
-          </div> 
-        </p:layoutUnit>
-      </p:layout>
-    </h:body>
-  </f:view>
-</html>
