--- hopsworks-kmon/src/main/webapp/monitor/host.xhtml
@@ -1,132 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Changes to this file committed after and not including commit-id: ccc0d2c5f9a5ac661e60e6eaf138de7889928b8b
-  ~ are released under the following license:
-  ~
-  ~ This file is part of Hopsworks
-  ~ Copyright (C) 2018, Logical Clocks AB. All rights reserved
-  ~
-  ~ Hopsworks is free software: you can redistribute it and/or modify it under the terms of
-  ~ the GNU Affero General Public License as published by the Free Software Foundation,
-  ~ either version 3 of the License, or (at your option) any later version.
-  ~
-  ~ Hopsworks is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
-  ~ without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
-  ~ PURPOSE.  See the GNU Affero General Public License for more details.
-  ~
-  ~ You should have received a copy of the GNU Affero General Public License along with this program.
-  ~ If not, see <https://www.gnu.org/licenses/>.
-  ~
-  ~ Changes to this file committed before and including commit-id: ccc0d2c5f9a5ac661e60e6eaf138de7889928b8b
-  ~ are released under the following license:
-  ~
-  ~ Copyright (C) 2013 - 2018, Logical Clocks AB and RISE SICS AB. All rights reserved
-  ~
-  ~ Permission is hereby granted, free of charge, to any person obtaining a copy of this
-  ~ software and associated documentation files (the "Software"), to deal in the Software
-  ~ without restriction, including without limitation the rights to use, copy, modify, merge,
-  ~ publish, distribute, sublicense, and/or sell copies of the Software, and to permit
-  ~ persons to whom the Software is furnished to do so, subject to the following conditions:
-  ~
-  ~ The above copyright notice and this permission notice shall be included in all copies or
-  ~ substantial portions of the Software.
-  ~
-  ~ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS  OR IMPLIED, INCLUDING
-  ~ BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-  ~ NONINFRINGEMENT. IN NO EVENT SHALL  THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
-  ~ DAMAGES OR  OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-  ~ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-  -->
-
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
-  "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml"   
-      xmlns:f="http://java.sun.com/jsf/core"
-      xmlns:h="http://java.sun.com/jsf/html"
-      xmlns:ui="http://java.sun.com/jsf/facelets"
-      xmlns:p="http://primefaces.org/ui"
-      >
-
-  <ui:composition template="layout.xhtml">
-
-    <ui:define name="title">
-      #{settingsController.name} - Host
-    </ui:define>
-
-    <ui:define name="content">
-
-      <!--Navigation Bar - Begin -->
-      <h:form styleClass="styleNav">
-        <p:panelGrid>  
-          <p:row>  
-            <p:column>
-              <p:commandLink id="hosts" action="hosts" ajax="false">  
-                <h:outputText value="Hosts" />
-              </p:commandLink> 
-              <h:outputText value=" &raquo; " />
-            </p:column>
-          </p:row>
-        </p:panelGrid>
-      </h:form>
-      <!--Navigation Bar - End-->
-
-      <p:outputPanel rendered="#{not hostController.found}">
-        <p:spacer width="1" height="50" />             
-        <p:panel header="Error" style="width: 300px; margin: 0 auto;">
-          <p align="center"> Host not found!</p>   
-        </p:panel>
-      </p:outputPanel>
-
-      <p:outputPanel rendered="#{hostController.found}">         
-        <!-- Host Toolbar -->
-        <h:form>  
-          <p:toolbar styleClass="SecondToolbar#{hostController.host.health}">  
-            <p:toolbarGroup align="left">  
-              <p:commandLink action="#{urlController.host()}" ajax="false">
-                <h:outputText value="#{hostController.host.hostname} (#{hostController.host.hostIp})"/>
-                <f:param name="hostname" value="#{hostController.hostname}"/>   
-              </p:commandLink>
-            </p:toolbarGroup>  
-
-          </p:toolbar>  
-        </h:form>
-
-        <p:spacer width="1" height="8" />    
-
-        <h:form styleClass="styleInfoPanel" id="form_host">
-
-          <p:spacer width="1" height="20" /> 
-
-          <p:panel header="Services">
-            <p:dataTable id="services" var="hostService" value="#{hostController.hostServices}">
-              <p:column headerText="Cluster"><h:outputText value="#{hostService.cluster}"/></p:column>
-              <p:column headerText="Group"><h:outputText value="#{hostService.group}"/></p:column>                            
-              <p:column headerText="Service"><h:outputText value="#{hostService.service}"/></p:column>                           
-              <p:column headerText="Service Status">
-                <p:commandLink action="#{urlController.serviceStatus}" ajax="false">
-                  <h:outputText value="#{hostService.host.hostname}"/>
-                  <f:param name="hostname" value="#{hostService.host.hostname}"/>
-                  <f:param name="cluster" value="#{hostService.cluster}"/>
-                  <f:param name="service" value="#{hostService.service}"/>
-                  <f:param name="group" value="#{hostService.group}"/>
-                </p:commandLink>                         
-              </p:column>  
-            </p:dataTable>                       
-          </p:panel>                     
-        </h:form>                                
-
-        <p:spacer width="1" height="40" />
-
-        <p:dialog modal="true" widgetVar="dlgAction" header="Action in progress..." draggable="false" closable="false" resizable="false">  
-          <p:graphicImage library="images" name="ajaxloadingbar.gif" />  
-        </p:dialog>
-
-        <p:spacer width="1" height="20" />
-      </p:outputPanel>
-
-    </ui:define>
-
-  </ui:composition>
-
-</html>
-
